Update patch set 1
Patch Set 1: Code-Review-2

I notice the following problems with this change:

* In the argouml repo, mvn clean test, fails. The problem is a build failure in argouml-app: [ERROR] Failed to execute goal org.jacoco:jacoco-maven-plugin:0.8.5:report (report) on project argouml: An error has occurred in JaCoCo report generation.: Error while creating report: Error while analyzing /home/linus/project/argouml/github@-repo/argouml/src/argouml-app/target/classes/build-eclipse/org/argouml/ui/MenuBar14.class. Can't add different class with same name: org/argouml/ui/MenuBar14 -> [Help 1]
This is strange since a change in the settings for the coverage report should not affect how the tests are run.

* In the argouml repo, mvn clean test site, fails with the same problem (in argouml app).

* In the argouml repo, mvn clean site, does not create the coverage report.

What is it you are trying to do with this chagne? Why does it affect the tests?

Patch-set: 1
Reviewer: Gerrit User 1016795 <1016795@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2