Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> I notice the following problems with this change:
> 
> * In the argouml repo, mvn clean test, fails. The problem is a build failure in argouml-app: [ERROR] Failed to execute goal org.jacoco:jacoco-maven-plugin:0.8.5:report (report) on project argouml: An error has occurred in JaCoCo report generation.: Error while creating report: Error while analyzing /home/linus/project/argouml/github@-repo/argouml/src/argouml-app/target/classes/build-eclipse/org/argouml/ui/MenuBar14.class. Can't add different class with same name: org/argouml/ui/MenuBar14 -> [Help 1]
> This is strange since a change in the settings for the coverage report should not affect how the tests are run.
> 
> * In the argouml repo, mvn clean test site, fails with the same problem (in argouml app).
> 
> * In the argouml repo, mvn clean site, does not create the coverage report.
> 
> What is it you are trying to do with this chagne? Why does it affect the tests?

Even don't know what to say.
It's something bad on your side...

mvn clean test

works for me

[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  13:13 min

Try to execute in the command line

repo sync
mvn clean test

it must work
Without it we shall not get the coverage
In the test phase the file target/jacoco.exec must be created
This file contains information about the coverage and then is used by site plugin.
Obviously if the file target/jacoco.exec is absent we shall not have coverage report.
What is it you are trying to do with this change?
I added jacoco coverage to be reported by site plugin in the section "Project Reports"
It really should not have influence on tests

mvn clean test site

also works for me

[INFO] BUILD SUCCESS
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  20:06 min

Patch-set: 1