Remove the configuration of the pmd tool
This will turn of the pmd tool and no such reports will be generated.

The reason for this is that we use a very old version of the tool.
That old version does not support the version 8 of java. The newer
version of the tool does not support version 6 or 7.

Change-Id: Id4ef075043c9900fd95637337a630a638a1aaafb